+++ mlgames/include/mlgames/AbstractGame.h
@@ -18,6 +18,6 @@ public:
     virtual size_t get_action_count() const = 0;
     virtual int get_score() const = 0;
     virtual StepResult get_step(int action) = 0;
-    virtual std::vector<int> get_state() const = 0;
+    virtual std::vector<double> get_state() const = 0;
     virtual void reset() = 0;
 };
